backend/requirements.txt
# =============================================================================
# AgriSense Backend Dependencies
# =============================================================================

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Shared utilities
numpy>=1.24.0

# AI Agents & LLM  
crewai>=0.28.0
langchain-groq>=0.0.1
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.0
litellm>=1.0.0
# Note: torch is installed separately in Dockerfile

# RAG & Vector Store
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
chromadb>=0.4.22
pypdf>=4.0.0
pyyaml>=6.0

# Reranking (Industry Standard — cross-encoder for precision)
sentence-transformers>=2.2.0

# ESP32-CAM Robotics Integration
onnxruntime>=1.16.0
opencv-python-headless>=4.8.0
httpx>=0.25.0
Pillow>=10.0.0
tensorflow>=2.15.0
websockets>=12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
